requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0
qrcode[pil]
flask-compress==1.13
